Allow MockSource to be used in other specs
